Add files via upload
I want to preface that I found a different library that will simplify this process much more I think, If you think that it's better to go with the MR library for the simulation, be my guest but I was having some major troubles getting the UR5.urdf file to open so I went in another direction and used the "Robotics Systems Toolbox" which can be added to your MATLAB for free.

So far with using this, I was only able to initialize an environment and the robot UR5 manipulator. This does look promising though. it seems like you are able to move the joints based on a configuration described in this link: https://www.mathworks.com/help/robotics/ref/rigidbodytree.html (This helped me a lot to understand what is going on with the code I have compiled here). 

What I would recommend we do next is to find a way to give an obstacle velocity so that we can show some results of our mission, or to find a way to get the manipulator to move from one position to another (possibly RRT). If you guys can figure any of this out, that would be extremely helpful. 

Thanks,
Connor